+++ src/main/java/shop/nuribooks/books/book/contributor/dto/ContributorRequest.java
@@ -5,7 +5,7 @@ import org.hibernate.validator.constraints.Length;
 import jakarta.validation.constraints.NotBlank;
 import lombok.Builder;
 
-import shop.nuribooks.books.book.contributor.entitiy.Contributor;
+import shop.nuribooks.books.book.contributor.entity.Contributor;
 
 @Builder
 public record ContributorRequest(@NotBlank @Length(min = 1, max = 50) String name) {
