+++ src/main/java/shop/nuribooks/books/book/contributor/entity/ContributorEditor.java
@@ -1,4 +1,4 @@
-package shop.nuribooks.books.book.contributor.entitiy;
+package shop.nuribooks.books.book.contributor.entity;
 
 import lombok.Builder;
 import lombok.Getter;
