+++ src/main/java/shop/nuribooks/books/book/contributor/repository/ContributorRepository.java
@@ -4,7 +4,7 @@ import java.util.Optional;
 
 import org.springframework.data.jpa.repository.JpaRepository;
 
-import shop.nuribooks.books.book.contributor.entitiy.Contributor;
+import shop.nuribooks.books.book.contributor.entity.Contributor;
 
 public interface ContributorRepository extends JpaRepository<Contributor, Long> {
 	Optional<Contributor> findByName(String name);
