+++ src/main/java/shop/nuribooks/books/book/contributor/repository/role/ContributorRoleRepository.java
@@ -4,8 +4,8 @@ import java.util.Optional;
 
 import org.springframework.data.jpa.repository.JpaRepository;
 
-import shop.nuribooks.books.book.contributor.entitiy.ContributorRole;
-import shop.nuribooks.books.book.contributor.entitiy.ContributorRoleEnum;
+import shop.nuribooks.books.book.contributor.entity.ContributorRole;
+import shop.nuribooks.books.book.contributor.entity.ContributorRoleEnum;
 
 public interface ContributorRoleRepository extends JpaRepository<ContributorRole, Long> {
 	Optional<ContributorRole> findByName(ContributorRoleEnum name);
