+++ src/main/java/shop/nuribooks/books/book/contributor/service/ContributorServiceImpl.java
@@ -6,8 +6,8 @@ import org.springframework.transaction.annotation.Transactional;
 import lombok.RequiredArgsConstructor;
 import shop.nuribooks.books.book.contributor.dto.ContributorRequest;
 import shop.nuribooks.books.book.contributor.dto.ContributorResponse;
-import shop.nuribooks.books.book.contributor.entitiy.Contributor;
-import shop.nuribooks.books.book.contributor.entitiy.ContributorEditor;
+import shop.nuribooks.books.book.contributor.entity.Contributor;
+import shop.nuribooks.books.book.contributor.entity.ContributorEditor;
 import shop.nuribooks.books.book.contributor.repository.ContributorRepository;
 import shop.nuribooks.books.exception.contributor.ContributorNotFoundException;
