+++ src/main/java/shop/nuribooks/books/book/contributor/service/role/ContributorRoleService.java
@@ -4,7 +4,7 @@ import java.util.List;
 
 import shop.nuribooks.books.book.contributor.dto.role.ContributorRoleRequest;
 import shop.nuribooks.books.book.contributor.dto.role.ContributorRoleResponse;
-import shop.nuribooks.books.book.contributor.entitiy.ContributorRole;
+import shop.nuribooks.books.book.contributor.entity.ContributorRole;
 
 public interface ContributorRoleService {
 	ContributorRoleResponse registerContributorRole(ContributorRoleRequest req);
