+++ src/main/java/shop/nuribooks/books/book/contributor/service/role/ContributorRoleServiceImpl.java
@@ -7,8 +7,8 @@ import org.springframework.stereotype.Service;
 import lombok.RequiredArgsConstructor;
 import shop.nuribooks.books.book.contributor.dto.role.ContributorRoleRequest;
 import shop.nuribooks.books.book.contributor.dto.role.ContributorRoleResponse;
-import shop.nuribooks.books.book.contributor.entitiy.ContributorRole;
-import shop.nuribooks.books.book.contributor.entitiy.ContributorRoleEnum;
+import shop.nuribooks.books.book.contributor.entity.ContributorRole;
+import shop.nuribooks.books.book.contributor.entity.ContributorRoleEnum;
 import shop.nuribooks.books.book.contributor.repository.role.ContributorRoleRepository;
 import shop.nuribooks.books.exception.contributor.ContributorRoleNotFoundException;
 import shop.nuribooks.books.exception.contributor.DuplicateEntityException;
