+++ src/test/java/shop/nuribooks/books/book/contributor/service/role/ContributorRoleServiceImplTest.java
@@ -15,8 +15,8 @@ import org.mockito.junit.jupiter.MockitoExtension;
 
 import shop.nuribooks.books.book.contributor.dto.role.ContributorRoleRequest;
 import shop.nuribooks.books.book.contributor.dto.role.ContributorRoleResponse;
-import shop.nuribooks.books.book.contributor.entitiy.ContributorRole;
-import shop.nuribooks.books.book.contributor.entitiy.ContributorRoleEnum;
+import shop.nuribooks.books.book.contributor.entity.ContributorRole;
+import shop.nuribooks.books.book.contributor.entity.ContributorRoleEnum;
 import shop.nuribooks.books.book.contributor.repository.role.ContributorRoleRepository;
 import shop.nuribooks.books.exception.contributor.ContributorRoleNotFoundException;
 import shop.nuribooks.books.exception.contributor.DuplicateEntityException;
